Do not compile with babel, just publish npm publish js
