+++ package.json
@@ -4,13 +4,9 @@
   "description": "make async/promise function execute only one at a time.",
   "main": "index.js",
   "scripts": {
-    "build": "run-s build:*",
-    "build:clean": "rm -rf lib/",
-    "build:babel": "babel src/ --out-dir lib/",
-    "pretest": "npm run build",
     "test": "run-s test:*",
     "test:standard": "standard",
-    "test:mocha": "NODE_ENV=test mocha 'test/*.js' -r @babel/polyfill -r @babel/register --exit",
+    "test:mocha": "NODE_ENV=test mocha 'test/*.js' --exit",
     "prepublishOnly": "npm test"
   },
   "keywords": [
@@ -23,19 +19,12 @@
   "author": "Sho Hashimoto <hashimoto@shokai.org>",
   "license": "MIT",
   "devDependencies": {
-    "@babel/cli": "^7.4.3",
-    "@babel/core": "^7.4.3",
-    "@babel/polyfill": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/register": "^7.4.0",
     "chai": "^4.2.0",
     "mocha": "^4.0.1",
     "npm-run-all": "^4.1.5",
     "standard": "^12.0.1"
   },
-  "dependencies": {},
   "directories": {
-    "lib": "lib",
     "test": "test"
   },
   "repository": {
