+++ test/delay.js
@@ -1,1 +1,3 @@
-export const delay = msec => new Promise(resolve => setTimeout(resolve, msec))
+const delay = msec => new Promise(resolve => setTimeout(resolve, msec))
+
+module.exports = { delay }
